+++ CHANGELOG.md
@@ -7,6 +7,10 @@ CHANGELOG
 ### Latest changes
 
 * VFS-8360 Fixed compatibility between Onezone GUI 21.02 and Oneprovider GUI 20.02 by moving shared properties caching to Oneprovider GUI
+* VFS-8038 Refactored build process to support faster builds on CI
+* VFS-8346 Disabled reattach action for detached datasets with deleted root file
+* VFS-8405 Added support for QoS requirements defined in hardlinks that were deleted
+* VFS-8233 Fixed unnecessary error message when removing detached dataset which source file was deleted
 * VFS-8452 Fixed showing correct error page when shared file has been deleted
 * VFS-7629 Added new datasets panel with archives browser in file browser
 * VFS-8379 Added auto-redirect to archives view after archive create